Extract an N3Index class
